fix: Use int32_t instead of long for std::bit_cast (cross-platform compatibility)
On Linux, long is 8 bytes (64-bit) while float is 4 bytes, causing
std::bit_cast to fail. Changed to int32_t which is always 4 bytes
on all platforms (Linux, Windows, macOS).

Fixes build error on Linux:
error: no matching function for call to 'bit_cast<long int>(float&)'

Related: Task #00018

+++ src/bwx_core/bwx_math.cpp
@@ -86,13 +86,13 @@ namespace bwx_sdk {
 
 		float bwxFastSqrt(float number)
 		{
-			long i;
+			int32_t i;
 			float x = number * 0.5F;
 			float y = number;
 
 			const float threehalfs = 1.5F;
 
-			i = std::bit_cast<long>(y);
+			i = std::bit_cast<int32_t>(y);
 			i = 0x5f3759df - (i >> 1);
 			y = std::bit_cast<float>(i);
 			y = y * (threehalfs - (x * y * y));
